# internal/api/router.go
// Package api is Chronicle's HTTP surface: the router every later endpoint
// hangs off, the two health probes the deploy path needs, and the credential
// surface from CHRN-71.
package api

import (
	"context"
	"encoding/json"
	"log/slog"
	"net/http"
	"time"

	"github.com/google/uuid"

	"github.com/Einlanzerous/chronicle/internal/api/wire"
	"github.com/Einlanzerous/chronicle/internal/audio"
	"github.com/Einlanzerous/chronicle/internal/store"
	"github.com/Einlanzerous/chronicle/internal/upload"
)

// Pinger is the slice of the store the readiness probe needs. An interface so
// readiness can be tested without a database.
type Pinger interface {
	Ping(ctx context.Context) error
}

// Accounts is the slice of the store the credential surface needs. Named here
// rather than taking *store.Store so the middleware can be tested against a
// fake — the assertion that an unauthenticated request reaches nothing should
// not need Postgres to run.
type Accounts interface {
	CreateUser(ctx context.Context, email, displayName, kind string) (store.User, error)
	GetUser(ctx context.Context, id uuid.UUID) (store.User, error)
	GetUserByEmail(ctx context.Context, email string) (store.User, error)
	UpdateDisplayName(ctx context.Context, id uuid.UUID, name string) (store.User, error)
	DeleteUser(ctx context.Context, id uuid.UUID) error
	ListMembers(ctx context.Context) ([]store.Member, error)

	MintToken(ctx context.Context, userID uuid.UUID, kind, label string, expiresAt *time.Time) (string, error)
	MintInvite(ctx context.Context, userID uuid.UUID, label string) (string, error)
	ReplaceDeviceInvite(ctx context.Context, userID uuid.UUID) (string, error)
	UserByToken(ctx context.Context, plaintext string) (store.User, error)
	RedeemInvite(ctx context.Context, plaintext, deviceLabel string) (store.User, string, error)
	RevokeToken(ctx context.Context, plaintext string) error
	ListSessions(ctx context.Context, userID uuid.UUID, currentPlaintext string) ([]store.Session, error)
	RevokeSession(ctx context.Context, userID, sessionID uuid.UUID, currentPlaintext string) (bool, error)
}

// Deps is what the router needs to serve.
type Deps struct {
	DB       Pinger
	Accounts Accounts
	Logger   *slog.Logger
	Version  string
	Commit   string

	// CFAccess verifies Cloudflare Access JWTs. Nil disables SSO, and
	// POST /auth/sso/cloudflare then answers sso_disabled rather than 404 — a
	// client can tell "not configured here" from "wrong URL".
	CFAccess *CFAccessVerifier

	// MobileBaseURL is the origin baked into an invite's sign-in link. Empty
	// omits the link, leaving clients on their own fallback.
	MobileBaseURL string

	// SecureCookies sets the Secure flag on the session cookie. It is a config
	// value rather than something derived from the request because TLS
	// terminates at Traefik: r.TLS is nil for every request this service ever
	// sees in the deployment it ships, WAN entrypoint included.
	SecureCookies bool

	// ProxySecret is the value Traefik stamps on X-Chronicle-Proxy-Secret, and
	// the whole of the trust decision for X-Forwarded-For when keying the
	// sign-in limiter (CHRN-75). Empty means believe nobody, which makes every
	// request through Traefik share one bucket.
	//
	// It is not authentication. Nothing outside clientIP may consult it.
	ProxySecret string

	// Audio is the on-disk store of recordings (CHRN-23). Nil until
	// CHRONICLE_AUDIO_DIR is set, and GET /admin/storage then answers 503
	// naming the variable rather than reporting a corpus of zero -- "no audio
	// configured" and "no audio yet" are different facts and must not render
	// the same.
	Audio *audio.Store

	// Corpus is the database side of that report.
	Corpus Corpus

	// Transcription backs GET /admin/transcription (CHRN-27). Nil answers 503
	// rather than reporting an empty corpus, for the reason Audio does.
	Transcription Transcription

	// Transcribing reports whether a transcription pump is actually running.
	// It is on the report because an operator reading a large `pending` count
	// otherwise cannot tell a backlog from a Chronicle that was never pointed
	// at an ASR service -- and those want completely different remedies.
	Transcribing bool

	// Triage is CHRN-33's batch surface. Nil answers 503 naming the variables
	// that turn it on, rather than 404 — "not configured here" and "wrong URL"
	// are different facts, and this is the same shape Audio and Transcription
	// already use.
	Triage Triage

	// Uploads is the direct ingest path (CHRN-20). Nil when there is no audio
	// store, in which case the four /memos/uploads routes answer 503 naming
	// CHRONICLE_AUDIO_DIR -- the same shape the storage report uses, and for
	// the same reason: "not configured here" and "wrong URL" are different
	// facts and a client should be able to tell them apart.
	Uploads *upload.Service
}

// api holds what the handlers share.
//
// It implements wire.ServerInterface, which is the compile-time half of
// CHRN-97's drift guard: an operation added to openapi.yaml with no method here
// fails `go build`, and a method whose operation is deleted fails it too.
var _ wire.ServerInterface = (*api)(nil)

type api struct {
	accounts      Accounts
	db            Pinger
	version       string
	commit        string
	logger        *slog.Logger
	cfAccess      *CFAccessVerifier
	mobileBaseURL string
	secureCookies bool
	proxySecret   string
	mismatch      *mismatchWarner
	absent        *mismatchWarner
	proxySeen     *proxySeen
	signInLimiter *ipRateLimiter
	audio         *audio.Store
	corpus        Corpus
	uploads       *upload.Service
	transcription Transcription
	transcribing  bool
	triage        Triage
}

// NewRouter builds the HTTP handler.
//
// Routes arrive two ways while CHRN-97 is in flight: the ones openapi.yaml
// describes are registered BY THE GENERATOR onto a policyRouter, and the rest
// are hand-registered on the same mux below, exactly as they were. Twenty of
// the twenty-six are now generated; the six that remain are the triage group,
// and they move with their payloads. A route that is in the document is never
// also in the second list, because the mux would refuse the duplicate pattern
// at boot.
//
// The two probes answer different questions and must not be collapsed:
//
//	/healthz — is this process alive? No dependencies. If it answers, the
//	           binary is running, and a restart is the wrong remedy for a
//	           database that is merely down.
//	/readyz  — can this process serve traffic? Pings the database. A load
//	           balancer takes an unready instance out of rotation; it does
//	           not kill it.
//
// They are also the only two routes reachable without a credential, with the
// two sign-in endpoints — which is now a property policy.go states and a test
// proves against the document, rather than one this comment asserts. Everything
// else requires a session; there is no unauthenticated read surface. /healthz
// staying dependency-free is load bearing beyond liveness: CHRN-59's QR
// onboarding probes it to check a server address before committing to it, which
// happens before any credential exists.
func NewRouter(d Deps) http.Handler {
	a := &api{
		accounts:      d.Accounts,
		db:            d.DB,
		version:       d.Version,
		commit:        d.Commit,
		logger:        d.Logger,
		cfAccess:      d.CFAccess,
		mobileBaseURL: d.MobileBaseURL,
		secureCookies: d.SecureCookies,
		proxySecret:   d.ProxySecret,
		mismatch:      &mismatchWarner{now: time.Now},
		absent:        &mismatchWarner{now: time.Now},
		proxySeen:     &proxySeen{},
		signInLimiter: newIPRateLimiter(signInRateWindow, signInRateBurst),
		audio:         d.Audio,
		corpus:        d.Corpus,
		uploads:       d.Uploads,
		transcription: d.Transcription,
		transcribing:  d.Transcribing,
		triage:        d.Triage,
	}

	mux := http.NewServeMux()

	// EVERY ROUTE IN openapi.yaml, REGISTERED BY THE GENERATOR (CHRN-97).
	//
	// Nobody writes these registrations, which is what makes the route set
	// unable to drift from the document. They land on a policyRouter rather
	// than on the mux directly, so each one is wrapped with the credential
	// policy.go declares for it -- and a route the document has and policy.go
	// does not is a panic here, at construction, rather than an open endpoint.
	//
	// ErrorHandlerFunc is supplied because its DEFAULT is
	// `http.Error(w, err.Error(), 400)`: text/plain, and leakier than the
	// handlers it would be answering for. See bindError.
	routed := newPolicyRouter(mux, a)
	wire.HandlerWithOptions(a, wire.StdHTTPServerOptions{
		BaseRouter:       routed,
		ErrorHandlerFunc: bindError(d.Logger),
	})

	if d.Accounts == nil {
		return requestLogger(d.Logger, mux)
	}

	// TRIAGE (CHRN-33) — the primary surface, and the one place derived state
	// becomes authored state.
	//
	// The last routes still registered by hand. Everything above them is in
	// openapi.yaml and was registered by the generator; these five and
	// GET /admin/triage move with their payloads, which are a domain model
	// rather than a report -- triage.BatchItem carries scribe.Proposal, and
	// transcribing CHRN-32's proposal contract into the document is its own
	// diff rather than a tail on this one.
	//
	// requireUser and not requireOwner: every account triages its own memos.
	// The author scoping is applied inside the service, PER ITEM on the POST as
	// well as on the GET, because a list that merely hides a memo is not access
	// control — a client naming an id directly never went through the list.
	mux.HandleFunc("GET /triage/batch", a.requireUser(a.handleTriageBatch))
	mux.HandleFunc("POST /triage/accept", a.requireUser(a.handleTriageAccept))

	// CHRN-34's two escapes. requireUser on the same grounds as the two above,
	// and scoped inside the service per memo: deferring somebody else's memo
	// answers exactly as deferring one that does not exist.
	mux.HandleFunc("POST /triage/hold", a.requireUser(a.handleTriageHold))
	mux.HandleFunc("POST /triage/release", a.requireUser(a.handleTriageRelease))
	mux.HandleFunc("GET /triage/deferred", a.requireUser(a.handleTriageDeferred))

	// What triage left behind: the backlog by age, and the decisions that did
	// not finish landing. Owner only, because it spans every author's corpus.
	mux.HandleFunc("GET /admin/triage", a.requireOwner(a.handleAdminTriage))

	return requestLogger(d.Logger, mux)
}

// GetHealthz answers the liveness probe. No dependencies, by design: a database
// that is down is not a reason to restart the binary, and CHRN-59's QR
// onboarding probes this before any credential exists.
func (a *api) GetHealthz(w http.ResponseWriter, r *http.Request) {
	body := wire.Health{Status: wire.Ok, Version: a.version}
	if a.commit != "" {
		body.Commit = &a.commit
	}
	writeJSON(w, http.StatusOK, body)
}

// GetReadyz answers the readiness probe. Pings the database, because a load
// balancer takes an unready instance out of rotation rather than killing it --
// which is the question liveness does not answer.
func (a *api) GetReadyz(w http.ResponseWriter, r *http.Request) {
	ctx, cancel := context.WithTimeout(r.Context(), 2*time.Second)
	defer cancel()
	if a.db != nil {
		if err := a.db.Ping(ctx); err != nil {
			a.logger.Warn("readiness probe failed", "check", "database", "error", err)
			check := "database"
			writeJSON(w, http.StatusServiceUnavailable, wire.Readiness{
				Status: wire.Unready,
				Check:  &check,
			})
			return
		}
	}
	writeJSON(w, http.StatusOK, wire.Readiness{Status: wire.Ready})
}

// serverError logs the cause and answers with a generic message. The detail
// belongs in the log, not in a response body that may cross the WAN.
func (a *api) serverError(w http.ResponseWriter, r *http.Request, what string, err error) {
	a.logger.ErrorContext(r.Context(), "request failed", "op", what, "error", err)
	writeError(w, http.StatusInternalServerError, codeInternal, "internal error")
}

// statusRecorder captures the status code for the access log.
type statusRecorder struct {
	http.ResponseWriter
	status int
}

func (r *statusRecorder) WriteHeader(code int) {
	r.status = code
	r.ResponseWriter.WriteHeader(code)
}

// requestLogger emits one structured line per request. Health probes log at
// debug so a 10-second liveness check does not bury everything else at info.
func requestLogger(logger *slog.Logger, next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		start := time.Now()
		rec := &statusRecorder{ResponseWriter: w, status: http.StatusOK}
		next.ServeHTTP(rec, r)

		level := slog.LevelInfo
		switch {
		case r.URL.Path == "/healthz" || r.URL.Path == "/readyz":
			level = slog.LevelDebug
		case rec.status >= 500:
			level = slog.LevelError
		case rec.status >= 400:
			level = slog.LevelWarn
		}
		// Datadog's standard log attributes, so these map without a custom
		// pipeline: http.method / http.status_code / http.url, and `duration`
		// in NANOSECONDS, which is what Datadog documents that field to be.
		// Dozzle shows the raw JSON, which stays readable either way.
		//
		// Note what is absent: no query string, no Authorization header, no
		// request body. A sign-in URL carries an invite in ?token=, and a log
		// line is exactly the wrong place for a live credential.
		logger.LogAttrs(r.Context(), level, "http request",
			slog.String("http.method", r.Method),
			slog.String("http.url", r.URL.Path),
			slog.Int("http.status_code", rec.status),
			slog.Int64("duration", time.Since(start).Nanoseconds()),
		)
	})
}

func writeJSON(w http.ResponseWriter, code int, body any) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(code)
	_ = json.NewEncoder(w).Encode(body)
}