# openapi.yaml
...
openapi: 3.0.3
info:
  title: Chronicle
  version: "1.0.0"
  description: |
    Voice-note ingestion into a notes and discussion wiki. A memo is captured,
    transcribed on the shared ASR service, routed by the Scribe into a note, a
    ticket or a discussion, and lands in a corpus that links out to Switchyard
    and Amber rather than absorbing them.

    **Two tiers, and they do not share a store.** Tier 1 is the estate's account
    of what exists — generated, regenerable, never hand-edited. Tier 2 is what a
    person said or wrote — memos, transcripts, notes, discussions — and none of
    it can be rebuilt. They are reached by different code paths through
    different database roles.

    **References are linked, never copied.** A ticket or an Amber item resolves
    at render time and is never written into Chronicle's tables. Resolution
    therefore has its own endpoint: a note payload carries the references its
    text names and nothing an upstream said about them.
servers:
  - url: http://chronicle:4009
    description: construct_net, container-to-container
  - url: https://chronicle.zerogravity.industries
    description: the Access-gated host

# Every operation needs a session unless it says otherwise. The two forms are
# the same credential: `sessionToken` reads the bearer header first and falls
# back to the cookie, so a client may present either.
security:
  - sessionBearer: []
  - sessionCookie: []
...
paths:
...
  /auth/sessions:
    get:
      tags: [auth]
      summary: Your own signed-in devices.
      description: |
        The one real risk in a password-free model: a session does not expire,
        so a lost device stays signed in forever unless its holder can see it
        and cut it off. Hence this list, and the revoke below it.
      operationId: listSessions
      x-chronicle-policy: member
      responses:
        "200":
          description: this account's sessions, the current one marked
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeviceSession"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalError" }
  /auth/sessions/{id}:
    delete:
      tags: [auth]
      summary: Revoke one of your own sessions.
      description: |
        Revoking the session you are calling with is a sign-out, and clears the
        cookie too. Another account's session answers exactly as one that does
        not exist.
      operationId: revokeSession
      x-chronicle-policy: member
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "204": { description: "revoked, or already gone" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }

  # ── ACCOUNTS ────────────────────────────────────────────────────────────────
  # Owner only, and never an agent. Not served on the WAN host at all:
  # construct-server's `chronicle-public-admin` router blackholes /admin there,
  # because POST /admin/users/{id}/invite mints a live credential and
  # requireOwner alone is not the boundary that should be holding it.
  /admin/users:
    get:
      tags: [accounts]
      summary: Every account, with its devices and invite state.
      operationId: listUsers
      x-chronicle-policy: owner
      responses:
        "200":
          description: the roster
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalError" }

    post:
      tags: [accounts]
      summary: Create an account and mint its first invite.
      operationId: createUser
      x-chronicle-policy: owner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewUserRequest"
      responses:
        "201":
          description: the account and its invite, shown once
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invite"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: that email is already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413": { $ref: "#/components/responses/TooLarge" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /admin/users/{id}:
    delete:
      tags: [accounts]
      summary: Remove an account and its credentials.
      description: |
        The owner cannot be removed, and neither can an author the corpus
        references — that answers `409`, because it is a refusal the caller can
        act on rather than a fault. Tier 2 is authored and irreplaceable, so a
        memo keeps the author that recorded it.
      operationId: deleteUser
      x-chronicle-policy: owner
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "204": { description: removed }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: the owner account cannot be removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: this account has memos and cannot be removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500": { $ref: "#/components/responses/InternalError" }

  # ── UPLOADS ─────────────────────────────────────────────────────────────────
  # Resumable memo ingest (CHRN-20). Every refusal here is a client-visible
  # answer with a specific remedy, which is why they are enumerated rather than
  # collapsed into 400: a phone deciding whether to resend, wait, or give up
  # needs to tell "we disagree about the offset" from "those were the wrong
  # bytes".
  #
  # `Upload-Offset` is the resume instruction and rides on the RESPONSE of every
  # state-carrying answer, including 409 and 408.
...
components:
  securitySchemes:
    # ONE credential, two ways of presenting it. `sessionToken` reads the
    # Authorization header first and falls back to the cookie, so an operation
    # accepting one accepts the other and the document says so with two
    # alternative requirements rather than by picking a favourite.
    sessionBearer:
      type: http
      scheme: bearer
      description: A session token minted by POST /auth/session or a redeemed invite.
    sessionCookie:
      type: apiKey
      in: cookie
      name: chronicle_session
      description: |
        The same token, set as an HttpOnly cookie by the sign-in routes. What
        the web client uses; an XSS cannot read it.
  parameters:
    SessionId:
      name: id
      in: path
      required: true
      description: A session's id, from GET /auth/sessions.
      schema:
        type: string
        format: uuid
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: |
        a parameter or a field is missing or malformed. The `code` carries the
        machine-readable part; the message stays flat rather than describing the
        value it rejected.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: |
        no such thing, for this caller. Another account's row answers this too —
        a list that merely hides a row is not access control.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TooLarge:
      description: the request body is larger than this endpoint accepts
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnsupportedMediaType:
      description: |
        the wrong `Content-Type`. On the JSON routes this is what closes login
        CSRF: a cross-site form cannot produce `application/json`, so the check
        is a security control rather than pedantry.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Unauthorized:
      description: no session, or one this deployment does not know
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: a valid session, on an account that may not do this
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: |
        something failed that the caller cannot act on. The detail is in the
        log and deliberately not in this body, which may cross the WAN.

        **Every credentialed operation declares this, and the reason is not
        that every handler has a bug.** `requireUser` resolves the session
        against the database, so a dropped connection answers here on ANY route
        behind a credential — the shared wrapper's 500, not the handler's. A
        document that omitted it would leave three generated clients with no
        branch for the answer they get when Postgres blinks.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: A stable machine token. Clients branch on this, never on `message`.
          example: owner_only
        message:
          type: string
          description: One sentence, for a person. Not stable across releases.
...
    NewUserRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          description: Also the Access identity this account is matched to on SSO.
        display_name:
          type: string
        kind:
          type: string
          description: |
            `person` or `agent`. An agent account can hold a token but is never
            an owner and never confirms authored text — `note_revisions_guard`
            refuses any revision whose `confirmed_by` names one.
    User:
      type: object
      description: |
        An account, as the wire carries it. It deliberately holds **no token
        material**: a credential is returned exactly once, by the call that
        mints it.
      required: [id, email, display_name, kind, is_owner]
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        display_name: { type: string }
        kind: { type: string }
        is_owner: { type: boolean }
    Member:
      description: |
        An account with what the owner needs to administer it. `last_seen_at`
        null and `invite_expires_at` set is an invite never redeemed.
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          required: [last_seen_at, invite_expires_at, session_count]
          properties:
            last_seen_at:
              type: string
              format: date-time
              nullable: true
            invite_expires_at:
              type: string
              format: date-time
              nullable: true
            session_count:
              type: integer
...
    Invite:
      type: object
      description: |
        A single-use invite, shown once. `sign_in_url` is the QR target and is
        absent when this deployment has no mobile base URL configured — a link
        built from the Access-gated host would walk a phone into an SSO wall a
        bearer token cannot open.
      required: [user, invite_token, expires_in]
      properties:
        user:
          $ref: "#/components/schemas/User"
        invite_token:
          type: string
        sign_in_url:
          type: string
        expires_in:
          type: string
          description: How long the invite is good for, as a Go duration string.
          example: 24h0m0s
    DeviceSession:
      type: object
      required: [id, device_label, created_at, last_seen_at, current]
      properties:
        id: { type: string, format: uuid }
        device_label: { type: string }
        created_at: { type: string, format: date-time }
        last_seen_at:
          type: string
          format: date-time
          nullable: true
        current:
          type: boolean
          description: |
            Whether this is the session making the request. What lets a person
            revoke every device but the one in their hand.
...